module refactor

+++ tripplebanner/sql/uninstall.php
@@ -1,15 +1,14 @@
 <?php
 /**
-*  @author    sHKamil - Kamil Hałasa
-*  @copyright sHKamil - Kamil Hałasa
-*  @license   .l
-*/
-
-/**
- * In some cases you should not drop the tables.
- * Maybe the merchant will just try to reset the module
- * but does not want to loose all of the data associated to the module.
+ *  @author    sHKamil - Kamil Hałasa
+ *  @copyright sHKamil - Kamil Hałasa
+ *   @license   GPL
  */
+
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
 $sql = [];
 $sql[] = "DROP TABLE IF EXISTS `" . _DB_PREFIX_ . "tripplebanner`";
